Separate code by build directive
httptrace is only available in go-1.7+
