+++ prometheus/promhttp/client_1_7_test.go
@@ -11,6 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +build go1.7
+
 package promhttp
 
 import (
